src/requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
pypdf
python-docx
python-pptx
sentence-transformers
faiss-cpu